Add Prepare a release - 022621.pdf document
This document describes the process of creating a ConnectBox software image. It includes instructions for making a build platform from either a Mac or a Raspberry Pi. It further describes the use of those platforms in compiling the ConnectBox source codes found on this GitHub site, into either a NanoPi NEO based ConnectBox image or a Raspberry Pi based ConnectBox image.